requirements.txt
questionary==2.0.1
pyyaml==6.0.1
rich==13.7.0
python-dotenv==1.0.0
flask==3.0.2
flask-cors==4.0.0 